Refactor effects and abilities

+++ Tests/GameExample/Attributes.cs
@@ -1,25 +1,26 @@
 using System;
 using HexCore.HexGraph;
+using NUnit.Framework.Constraints;
 
 namespace Tests.GameExample
 {
     [Serializable]
     public class Attributes
     {
-        public int Attack;
+        public double Attack;
         public int AttackRange;
         public bool CanAttack;
         public bool CanCast;
         public bool CanMove;
-        public int Defense;
-        public int HP;
-        public int MagicPower;
+        public double Defense;
+        public double HP;
+        public double MagicPower;
         public int MovementRange;
         public MovementType MovementType;
-        public int MP;
+        public double MP;
 
-        public Attributes(MovementType movementType = null, int attack = 0, int defense = 0, int hp = 0, int mp = 0,
-            int magicPower = 0, int movementRange = 0, int attackRange = 0)
+        public Attributes(MovementType movementType = null, double attack = 0, double defense = 0, double hp = 0,
+            double mp = 0, double magicPower = 0, int movementRange = 0, int attackRange = 0)
         {
             Attack = attack;
             Defense = defense;
